backend/requirements.txt
# FastAPI and ASGI server
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.18

# Database
sqlalchemy==2.0.36
alembic==1.14.0
psycopg2-binary==2.9.10
asyncpg==0.30.0

# Pydantic
pydantic==2.10.3
pydantic-settings==2.6.1

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1

# AI Services
openai==1.58.1
anthropic==0.39.0
httpx==0.28.1

# Utilities
python-dateutil==2.9.0
